rfm_backend/requirements.txt
# Requirements for Adheseal RFM Analysis Dashboard Backend
fastapi==0.92.0
uvicorn==0.20.0
pandas==2.0.0
pydantic==1.10.5
python-dotenv==1.0.0
pytest==7.3.1
httpx==0.27.0
